feat(libft/printf): add support for size specifiers

+++ libft/io/printf/ft_printf.h
@@ -5,8 +5,8 @@
 # include <libft/string.h>
 # include <stdarg.h>
 
-int	printf_convert_int(t_string *buffer, int n);
-int	printf_convert_unsigned(t_string *buffer, unsigned int n);
+int	printf_convert_int(t_string *buffer, int64_t n);
+int	printf_convert_unsigned(t_string *buffer, uint64_t n);
 int	printf_convert_hex(t_string *buffer, unsigned int n);
 int	printf_convert_hex_upper(t_string *buffer, unsigned int n);
 int	printf_convert_pointer(t_string *buffer, void *ptr);
